+++ src/lib/sys.h
@@ -241,7 +241,7 @@ long sys_file_hardlink_count(const char *file);
  *
  * @return Return the size of a file or a negative number on errors.
  */
-long sys_get_file_size(const char *filename);
+ssize_t sys_get_file_size(const char *filename);
 
 /**
  * @brief Run a systemctl command with the informed parameters.
